mm-video: hevc: initialize ion heap_mask for hevc
The heap_mask was not initialized correctly and contained
garbage values. This resulted in buffers being allocated
from vmalloc heap instead of adsp heap. ion_phys is not
supported for vmalloc heap and this resulted in prepare_buf
call failing as driver was unable to set the physical
address of output buffers to Q6 firmware.

Change-Id: Id4a55a7d6b7f0e6b03e33fd9bc9cd6d1e363f70b